:truck: move containers/Layout back to components/layout/Layout
